Change PM2 management
From now on, no processes are removed from PM2 ls when one stops an instance.
The instance.json is modified just to have its turnoffAll value toggled from true to false or the reverse.
As PM2 watches the instance.json, it will restart the instance, which will be running, but not doing anything.